Allow measurement of non-numerical quantities (e.g., spectral type)

+++ bioverse/survey.py
@@ -339,8 +339,9 @@ class Measurement():
 
         # Place this measurement into the measurements database
         if self.key not in data.keys():
-            data[self.key] = np.full(len(detected), np.nan)
-            data.error[self.key] = np.full(len(detected), np.nan)
+            data[self.key] = np.full(len(detected), np.nan, dtype=x.dtype)
+            data.error[self.key] = np.full(len(detected), np.nan, dtype=x.dtype)
+
         data[self.key][observable] = x
         data.error[self.key][observable] = dx
